chore: move « guessRenderer » into core folder
