+++ lib/core/guessRenderer.js
